Revert "Update git comment character"
This reverts commit 607930dc1fccf0fb391fb04fc18a80b582736343.

Realized we're using tree-sitter for parsing git commit messages,
which is [hard-coded][0] to use '#' as the comment character. Would rather
temporarily set `-c core.commentChar=;` to commit command than try
to patch the parser. Could also just uninstall the git commit parser,
but it's nice to just stick with tree-sitter for everything.

The [default syntax file][1] does runtime matching of the comment
character, which is probably not possible for a tree-sitter parser?
Haven't actually checked though.

[0]: https://github.com/gbprod/tree-sitter-gitcommit/blob/edd817e0532f179b7f7f371dc180629070945f0c/src/grammar.json#L402-L455
[1]: https://github.com/neovim/neovim/blob/cdd87222c86c5b2274a13d36f23de0637462e317/runtime/syntax/gitcommit.vim#L40