Get nvidia driver + external monitor to work without flickering
This also assumes that the BIOS is configured to boot with only
the discrete GPU (dGPU). Otherwise, the builtin display is driven by the
integrated GPU (iGPU), and the external monitor is driven by the dGPU,
which apparently leads to horrendous flickering. Booting without the
dGPU to conserve battery will require changing the BIOS setting :(

We don't need PRIME synchronization in this setup since all displays
are being rendered by the dGPU.

Also, disabling the iGPU caused /sys/class/backlight to break,
but using the native acpi_backlight device evidently works.

Change the corresponding kanshi criteria since the BIOS setting
causes the sway outputs to be renamed.