Fix wezterm after system upgrade
Some combination of OpenGL/defaulting to integrated graphics/Wayland
vs XWayland was causing only blocks of color to be displayed, rather
than text. Too tired of debugging these kinds of graphics issues
to figure out root cause.

On the bright side, we can get rid of xwayland and use the Nvidia
GPU for rendering now, instead of the software backend.